build: Simple debian packaging for quilt
This patch adds functionality to create debian packaging for quilt
(i.e. a .deb file that can be installed easily by users). Output is
generated within the dist/ directory.

It does not completely comply with debian packaging policy as we
use dependencies that are not in debian yet, and the package is
currently not run through lintian. It should be sufficient for
test distribution for the time being.