docs: Add versioning information to minion section
When developing the minion, our current rule in the Makefile does not tag the built minion as "master". When developers write minion code and use the rules in the Makefile to create their own minion image, the deployed minions will try to pull an image tagged with "master" which won't exist. This patch instructs developers to change the version to "latest" for local development so that their personal minions will try to pull the "latest" tag of their build image.